backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database
sqlalchemy==2.0.35
asyncpg==0.29.0
alembic==1.13.3
geoalchemy2==0.15.2
psycopg2-binary==2.9.9

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12

# HTTP client
httpx==0.27.2

# Cache & Queue
redis==5.1.1
celery==5.4.0

# Validation
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic[email]==2.9.2
email-validator==2.2.0

# S3
boto3==1.35.23

# Utils
python-dotenv==1.0.1
shapely==2.0.6

# Tests
pytest==8.3.3
pytest-asyncio==0.24.0
anyio==4.6.0